i_overlay::core::graph

Struct OverlayGraph

pub struct OverlayGraph<'a> { /* private fields */ }

A representation of geometric shapes organized for efficient boolean operations.

OverlayGraph is a core structure designed to facilitate the execution of boolean operations on shapes, such as union, intersection, and difference. It organizes and preprocesses geometric data, making it optimized for these operations. This struct is the result of compiling shape data into a form where boolean operations can be applied directly, efficiently managing the complex relationships between different geometric entities.

Use OverlayGraph to perform boolean operations on the geometric shapes you’ve added to an Overlay, after it has processed the shapes according to the specified build and overlay rules. More information about Overlay Graph.

Implementations

impl OverlayGraph<'_>

pub fn extract_shapes( &self, overlay_rule: OverlayRule, buffer: &mut BooleanExtractionBuffer, ) -> IntShapes

Extracts shapes from the overlay graph based on the specified overlay rule. This method is used to retrieve the final geometric shapes after boolean operations have been applied. It’s suitable for most use cases where the minimum area of shapes is not a concern.

overlay_rule: The boolean operation rule to apply when extracting shapes from the graph, such as union or intersection.

buffer: Reusable buffer, optimisation purpose only.

Returns: A vector of IntShape, representing the geometric result of the applied overlay rule.

Shape Representation

The output is a IntShapes, where:

The outer Vec<IntShape> represents a set of shapes.

Each shape Vec<IntContour> represents a collection of contours, where the first contour is the outer boundary, and all subsequent contours are holes in this boundary.

Each path Vec<IntPoint> is a sequence of points, forming a closed path.

Note: Outer boundary paths have a counterclockwise order, and holes have a clockwise order.

pub fn extract_contours_into( &self, overlay_rule: OverlayRule, buffer: &mut BooleanExtractionBuffer, output: &mut FlatContoursBuffer, )

Extracts the flat contours from the overlay graph based on the specified overlay rule.

This method performs a Boolean operation (e.g., union or intersection) and stores the result directly into a flat buffer of contours, without nesting them into shapes (i.e., no hole-joining or grouping).

It is optimized for performance and suitable when raw contour data is sufficient, such as during intermediate processing, visualization, or tesselation.

overlay_rule: The boolean operation rule to apply (e.g., union, intersection, xor).

buffer: Reusable working buffer to avoid reallocations.

output: A flat buffer to which the resulting valid contours will be written.

impl OverlayGraph<'_>

pub fn validate(&self)

impl OverlayGraph<'_>

pub fn extract_separate_vectors(&self) -> Vec<VectorEdge>

pub fn extract_shape_vectors( &self, overlay_rule: OverlayRule, buffer: &mut BooleanExtractionBuffer, ) -> Vec<VectorShape>