i_overlay::float::string_graph

Struct FloatStringGraph

pub struct FloatStringGraph<'a, P: FloatPointCompatible<T>, T: FloatNumber> {
    pub graph: StringGraph<'a>,
    pub adapter: FloatPointAdapter<P, T>,
}

The FloatStringGraph struct represents a graph structure with floating-point precision, providing methods to extract geometric shapes from the graph after applying string-based operations.

Fields
graph: StringGraph<'a>adapter: FloatPointAdapter<P, T>
Implementations

impl<P: FloatPointCompatible<T>, T: FloatNumber> FloatStringGraph<'_, P, T>

pub fn extract_shapes(&self, string_rule: StringRule) -> Shapes<P>

Extracts shapes from the overlay graph based on the specified string rule. This method is used to retrieve the final geometric shapes after boolean operations have been applied. It’s suitable for most use cases where the minimum area of shapes is not a concern.

Parameters

string_rule: The string operation rule to apply when extracting shapes from the graph, such as slice.

Returns

A Shapes<P> collection, representing the geometric result of the applied string rule.

Shape Representation

The output is a Shapes<P>, where:

The outer Shapes<P> represents a set of shapes.

Each shape Shape<P> represents a collection of paths, where the first path is the outer boundary, and all subsequent paths are holes in this boundary.

Each path Contour<P> is a sequence of points, forming a closed path.

Note: Outer boundary paths have a counterclockwise order, and holes have a clockwise order.

pub fn extract_shapes_custom( &self, string_rule: StringRule, options: OverlayOptions<T>, ) -> Shapes<P>

Extracts shapes from the overlay graph similar to extract_shapes, but with an additional constraint on the minimum area of the shapes. This is useful for filtering out shapes that do not meet a certain size threshold, which can be beneficial for eliminating artifacts or noise from the output.

Parameters

string_rule: The string operation rule to apply when extracting shapes from the graph, such as slice.

options: Adjust custom behavior.

Returns

A Shapes<P> collection, representing the geometric result of the applied string rule.

Shape Representation

The output is a Shapes<P>, where:

The outer Shapes<P> represents a set of shapes.

Each shape Shape<P> represents a collection of paths, where the first path is the outer boundary, and all subsequent paths are holes in this boundary.

Each path Contour<P> is a sequence of points, forming a closed path.

Note: Outer boundary paths have a main_direction order, and holes have an opposite to main_direction order.