i_overlay::mesh::style

Struct StrokeStyle

pub struct StrokeStyle<P: FloatPointCompatible<T>, T: FloatNumber> {
    pub width: T,
    pub start_cap: LineCap<P, T>,
    pub end_cap: LineCap<P, T>,
    pub join: LineJoin<T>,
}

Defines the stroke style for outlining paths.

Fields
width: T

The width of the stroke.

start_cap: LineCap<P, T>

The cap style at the start of the stroke.

end_cap: LineCap<P, T>

The cap style at the end of the stroke.

join: LineJoin<T>

The join style where two lines meet.

Implementations

impl<P: FloatPointCompatible<T>, T: FloatNumber> StrokeStyle<P, T>

pub fn new(width: T) -> Self

Creates a new StrokeStyle with the specified width.

pub fn width(self, width: T) -> Self

Sets the stroke width.

pub fn start_cap(self, cap: LineCap<P, T>) -> Self

Sets the cap style at the start of the stroke.

pub fn end_cap(self, cap: LineCap<P, T>) -> Self

Sets the cap style at the end of the stroke.

pub fn line_join(self, join: LineJoin<T>) -> Self

Sets the line join style.