i_overlay::string::clip

Trait IntClip

pub trait IntClip {
    // Required methods
    fn clip_line(
        &self,
        line: IntLine,
        fill_rule: FillRule,
        clip_rule: ClipRule,
    ) -> Vec<IntPath>;
    fn clip_lines(
        &self,
        lines: &[IntLine],
        fill_rule: FillRule,
        clip_rule: ClipRule,
    ) -> Vec<IntPath>;
    fn clip_path(
        &self,
        path: &IntPath,
        fill_rule: FillRule,
        clip_rule: ClipRule,
    ) -> Vec<IntPath>;
    fn clip_paths(
        &self,
        paths: &[IntPath],
        fill_rule: FillRule,
        clip_rule: ClipRule,
    ) -> Vec<IntPath>;
}

Required Methods

fn clip_line( &self, line: IntLine, fill_rule: FillRule, clip_rule: ClipRule, ) -> Vec<IntPath>

Clips a single line according to the specified build and clip rules.

line: The line to be clipped, represented by two points.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of line segments.

clip_rule: The rule for clipping, determining how the boundary and inversion settings affect the result.

Returns

A vector of IntPath instances representing the clipped sections of the input line.

fn clip_lines( &self, lines: &[IntLine], fill_rule: FillRule, clip_rule: ClipRule, ) -> Vec<IntPath>

Clips multiple lines according to the specified build and clip rules.

lines: A slice of IntLine instances representing lines to be clipped.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of line segments.

clip_rule: The rule for clipping, determining how boundary and inversion settings affect the results.

Returns

A vector of IntPath instances containing the clipped portions of the input lines.

fn clip_path( &self, path: &IntPath, fill_rule: FillRule, clip_rule: ClipRule, ) -> Vec<IntPath>

Clips a single path according to the specified build and clip rules.

path: A reference to an IntPath, which is a sequence of points representing the path to be clipped.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of path segments.

clip_rule: The rule for clipping, determining how boundary and inversion settings affect the result.

Returns

A vector of IntPath instances representing the clipped sections of the path.

fn clip_paths( &self, paths: &[IntPath], fill_rule: FillRule, clip_rule: ClipRule, ) -> Vec<IntPath>

Clips multiple paths according to the specified build and clip rules.

paths: A slice of IntPath instances, each representing a path to be clipped.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of path segments.

clip_rule: The rule for clipping, determining how boundary and inversion settings affect the result.

Returns

A vector of IntPath instances containing the clipped portions of the input paths.