i_overlay::float::simplify

Trait SimplifyShape

pub trait SimplifyShape<P, T: FloatNumber> {
    // Required methods
    fn simplify_shape(&self, fill_rule: FillRule) -> Shapes<P>;
    fn simplify_shape_custom(
        &self,
        fill_rule: FillRule,
        options: OverlayOptions<T>,
        solver: Solver,
    ) -> Shapes<P>;
}

Trait Simplify provides a method to simplify geometric shapes by reducing the number of points in contours or shapes while preserving overall shape and topology. The method applies a minimum area threshold and a build rule to determine which areas should be retained or excluded.

Required Methods

fn simplify_shape(&self, fill_rule: FillRule) -> Shapes<P>

Simplifies the shape or collection of points, contours, or shapes, based on a specified minimum area threshold.

Returns: A collection of Shapes<P> that represents the simplified geometry.

Note: Outer boundary paths have a main_direction order, and holes have an opposite to main_direction order.

fn simplify_shape_custom( &self, fill_rule: FillRule, options: OverlayOptions<T>, solver: Solver, ) -> Shapes<P>

Simplifies the shape or collection of points, contours, or shapes, based on a specified minimum area threshold.

options: Adjust custom behavior.

solver: Type of solver to use.

Returns: A collection of Shapes
that represents the simplified geometry.

Note: Outer boundary paths have a main_direction order, and holes have an opposite to main_direction order.