i_overlay::mesh::outline::offset

Trait OutlineOffset

pub trait OutlineOffset<P: FloatPointCompatible<T>, T: FloatNumber> {
    // Required methods
    fn outline(&self, style: &OutlineStyle<T>) -> Shapes<P>;
    fn outline_custom(
        &self,
        style: &OutlineStyle<T>,
        options: OverlayOptions<T>,
    ) -> Shapes<P>;
}

Required Methods

fn outline(&self, style: &OutlineStyle<T>) -> Shapes<P>

Generates an outline shapes for contours, or shapes.

style: Defines the outline properties, including offset, and joins.

Returns

A collection of Shapes<P> representing the outline geometry. Note: Outer boundary paths have a counterclockwise order, and holes have a clockwise order.

fn outline_custom( &self, style: &OutlineStyle<T>, options: OverlayOptions<T>, ) -> Shapes<P>

Generates an outline shapes for contours, or shapes with optional filtering.

style: Defines the outline properties, including offset, and joins.

options: Adjust custom behavior.

Returns

A collection of Shapes<P> representing the outline geometry. Note: Outer boundary paths have a main_direction order, and holes have an opposite to main_direction order.