i_overlay::string::overlay

Struct StringOverlay

pub struct StringOverlay {
    pub options: IntOverlayOptions,
    /* private fields */
}

Fields
options: IntOverlayOptions
Implementations

impl StringOverlay

pub fn new(capacity: usize) -> Self

Constructs a new StringOverlay instance, initializing it with a capacity that should closely match the total count of edges from all shapes being processed. This pre-allocation helps in optimizing memory usage and performance.

capacity: The initial capacity for storing edge data. Ideally, this should be set to the sum of the edges of all shapes to be added to the overlay, ensuring efficient data management.

pub fn with_options(capacity: usize, options: IntOverlayOptions) -> Self

Constructs a new StringOverlay instance, initializing it with a capacity that should closely match the total count of edges from all shapes being processed. This pre-allocation helps in optimizing memory usage and performance.

capacity: The initial capacity for storing edge data. Ideally, this should be set to the sum of the edges of all shapes to be added to the overlay, ensuring efficient data management.

options: Adjust custom behavior.

pub fn with_shape_contour(contour: &[IntPoint]) -> Self

Creates a new StringOverlay instance and initializes it with a single shape contour.

contour: An array of points that form a closed path.

pub fn with_shape_contours(contours: &[IntContour]) -> Self

Creates a new StringOverlay instance and initializes it with multiple shape contours.

contours: An array of IntContour instances to be added to the overlay.

pub fn with_shape(shape: &[IntContour]) -> Self

Creates a new StringOverlay instance and initializes it with s shape.

shape: An IntShape instances to be added to the overlay.

pub fn with_shapes(shapes: &[IntShape]) -> Self

Creates a new StringOverlay instance and initializes it with subject and clip shapes.

shapes: An array of IntShape instances to be added to the overlay.

pub fn add_shape_contour_iter<I: Iterator<Item = IntPoint>>(&mut self, iter: I)

Adds a path to the overlay using an iterator, allowing for more flexible path input. This function is particularly useful when working with dynamically generated paths or when paths are not directly stored in a collection.

iter: An iterator over references to IntPoint that defines the path.

pub fn add_shape_contour(&mut self, contour: &[IntPoint])

Adds a single path to the overlay as a shape paths.

contour: An array of points that form a closed path.

pub fn add_shape_contours(&mut self, contours: &[IntContour])

Adds multiple paths to the overlay as shape paths.

contours: An array of IntContour instances to be added to the overlay.

pub fn add_shapes(&mut self, shapes: &[IntShape])

Adds a list of shape to the overlay.

shapes: An array of IntShape instances to be added to the overlay.

pub fn add_string_line(&mut self, line: IntLine)

Adds a single line (open path) to the overlay.

line: An IntLine representing the open line (defined by two points).

pub fn add_string_lines(&mut self, lines: &[IntLine])

Adds multiple lines (open paths) to the overlay.

lines: An array of IntLine instances to be added.

pub fn add_string_path(&mut self, path: &[IntPoint])

Adds a string path to the overlay.

path: A path representing a string line.

pub fn add_string_contour(&mut self, contour: &[IntPoint])

Adds a string line contour to the overlay.

contour: A contour representing a string line closed path. This path is interpreted as closed, so it doesn’t require the start and endpoint to be the same for processing.

pub fn add_string_paths(&mut self, paths: &[IntPath])

Adds a string line paths to the overlay.

paths: A collection of paths, each representing a string line.

pub fn add_string_contours(&mut self, contours: &[IntContour])

Adds a string line contours to the overlay.

contours: A collection of contours, each representing a string line closed path.

pub fn clip_string_lines( self, fill_rule: FillRule, clip_rule: ClipRule, ) -> Vec<IntPath>

Clips lines according to the specified build and clip rules.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of line segments.

clip_rule: The rule for clipping, determining how the boundary and inversion settings affect the result.

Returns

A vector of IntPath instances representing the clipped sections of the input lines.

pub fn clip_string_lines_with_solver( self, fill_rule: FillRule, clip_rule: ClipRule, solver: Solver, ) -> Vec<IntPath>

Clips lines according to the specified build and clip rules.

fill_rule: Specifies the rule determining the filled areas, influencing the inclusion of line segments.

clip_rule: The rule for clipping, determining how the boundary and inversion settings affect the result.

solver: A solver type to be used for advanced control over the graph building process.

Returns

A vector of IntPath instances representing the clipped sections of the input lines.

pub fn build_graph_view( &mut self, fill_rule: FillRule, ) -> Option<StringGraph<'_>>

Builds and returns a lightweight, borrowed view of the overlay graph. This graph is used for string operations, enabling analysis and manipulation of geometric data.

fill_rule: The rule that defines how to build shapes (e.g., non-zero, even-odd).

pub fn build_graph_view_with_solver( &mut self, fill_rule: FillRule, solver: Solver, ) -> Option<StringGraph<'_>>

Builds and returns a lightweight, borrowed view of the overlay graph. This graph is used for string operations, enabling analysis and manipulation of geometric data.

fill_rule: The rule that defines how to build shapes (e.g., non-zero, even-odd).

solver: A solver type to be used for advanced control over the graph building process.